+++ tamr_unify_client/models/dataset/resource.py
@@ -58,7 +58,7 @@ class Dataset(BaseResource):
         """Send a batch of record creations/updates/deletions to this dataset.
 
         :param records: Each record should be formatted as specified in the `Public Docs for Dataset updates <https://docs.tamr.com/reference#modify-a-datasets-records>`_.
-        :type records: list[dict]
+        :type records: iterable[dict]
         """
         body = "\n".join([json.dumps(r) for r in records])
         self.client.post(self.api_path + ":updateRecords", data=body)
@@ -92,6 +92,36 @@ class Dataset(BaseResource):
         status_json = self.client.get(self.api_path + "/status").successful().json()
         return DatasetStatus.from_json(
             self.client, status_json, api_path=self.api_path + "/status"
+        )
+
+    def from_geo_features(self, features):
+        """Upsert this dataset from a geospatial FeatureCollection or iterable of Features.
+
+        `features` can be:
+
+        - An object that implements ``__geo_interface__`` as a FeatureCollection
+          (see https://gist.github.com/sgillies/2217756)
+        - An iterable of features, where each element is a feature dictionary or an object
+          that implements the ``__geo_interface__`` as a Feature
+        - A map where the "features" key contains an iterable of features
+
+        See: geopandas.GeoDataFrame.from_features()
+
+        :param features: geospatial features
+        """
+        if hasattr(features, "__geo_interface__"):
+            features = features.__geo_interface__
+        if hasattr(features, "get") and features.get("type") == "FeatureCollection":
+            features = features["features"]
+
+        key_attrs = self.key_attribute_names
+        if len(key_attrs) == 1:
+            record_id = "recordId"
+        else:
+            record_id = "compositeRecordId"
+
+        self.update_records(
+            self._features_to_updates(features, record_id, key_attrs, self._geo_attr)
         )
 
     @property
@@ -129,20 +159,22 @@ class Dataset(BaseResource):
             def key_value(rec):
                 return [rec[attr] for attr in key_attrs]
 
+        for record in self.records():
+            yield self._record_to_feature(record, key_value, key_attrs, self._geo_attr)
+
+    @property
+    def _geo_attr(self):
+        """The name of the attribute that contains geometry
+
+        :return: the name of the attribute that contains geometry
+        :rtype: str
+        """
         # Duck-typing: find all the attributes that look like geometry
-        geo_attr_names = {
-            "point",
-            "multiPoint",
-            "lineString",
-            "multiLineString",
-            "polygon",
-            "multiPolygon",
-        }
         geo_attrs = [
             attr.name
             for attr in self.attributes
             if "RECORD" == attr.type.base_type
-            and geo_attr_names.intersection(
+            and self._geo_attr_names().intersection(
                 {sub_attr.name for sub_attr in attr.type.attributes}
             )
         ]
@@ -151,8 +183,7 @@ class Dataset(BaseResource):
             geo_attr = geo_attrs[0]
         else:
             geo_attr = None
-        for record in self.records():
-            yield self._record_to_feature(record, key_value, key_attrs, geo_attr)
+        return geo_attr
 
     @staticmethod
     def _record_to_feature(record, key_value, key_attrs, geo_attr):
@@ -166,20 +197,13 @@ class Dataset(BaseResource):
         """
         feature = {"type": "Feature", "id": key_value(record)}
         reserved = {"bbox", geo_attr}.union(key_attrs)
-        conversion = {
-            "point": "Point",
-            "multiPoint": "MultiPoint",
-            "lineString": "LineString",
-            "multiLineString": "MultiLineString",
-            "polygon": "Polygon",
-            "multiPolygon": "MultiPolygon",
-        }
         if geo_attr and geo_attr in record:
             src_geo = record[geo_attr]
-            for unify_attr in conversion.keys():
+            for unify_attr in Dataset._geo_attr_names():
                 if unify_attr in src_geo and src_geo[unify_attr]:
                     feature["geometry"] = {
-                        "type": conversion[unify_attr],
+                        # Convert e.g. multiLineString -> MultiLineString
+                        "type": unify_attr[0].upper() + unify_attr[1:],
                         "coordinates": src_geo[unify_attr],
                     }
                     break
@@ -190,6 +214,56 @@ class Dataset(BaseResource):
             feature["properties"] = {attr: record[attr] for attr in non_reserved}
         return feature
 
+    @staticmethod
+    def _feature_to_record(feature, key_attrs, geo_attr):
+        """Convert a Python Geo Interface Feature to a Unify record
+
+        feature can be a dict representing a Geospatial Feature, or a Feature object
+        that implements the __geo_interface__ property.
+
+        :param feature: Python Geo Interface Feature
+        :param key_attrs: Sequence of attributes that comprise the primary key for the record
+        :param geo_attr: The singluar attribute on the record to use for the geometry
+        :return: dict
+        """
+        if hasattr(feature, "__geo_interface__"):
+            feature = feature.__geo_interface__
+
+        record = {}
+
+        props = feature.get("properties")
+        if props:
+            for prop in props:
+                record[prop] = props[prop]
+
+        geometry = feature.get("geometry")
+        if geometry:
+            geo_type = geometry["type"]
+            # Convert e.g. "MultiLineString" -> "multiLineString"
+            geo_type = geo_type[0].lower() + geo_type[1:]
+            record[geo_attr] = {geo_type: geometry["coordinates"]}
+
+        bbox = feature.get("bbox")
+        if bbox:
+            record["bbox"] = bbox
+
+        if key_attrs[1:]:
+            key_values = feature["id"]
+            for i, attr in enumerate(key_attrs):
+                record[attr] = key_values[i]
+        else:
+            record[key_attrs[0]] = feature["id"]
+        return record
+
+    @staticmethod
+    def _features_to_updates(features, id_attr, key_attrs, geo_attr):
+        for feature in features:
+            yield {
+                "action": "CREATE",
+                id_attr: feature["id"],
+                "record": Dataset._feature_to_record(feature, key_attrs, geo_attr),
+            }
+
     def __repr__(self):
         return (
             f"{self.__class__.__module__}."
@@ -198,3 +272,14 @@ class Dataset(BaseResource):
             f"name={self.name!r}, "
             f"version={self.version!r})"
         )
+
+    @staticmethod
+    def _geo_attr_names():
+        return {
+            "point",
+            "multiPoint",
+            "lineString",
+            "multiLineString",
+            "polygon",
+            "multiPolygon",
+        }