+++ tests/unit/test_dataset_geo.py
@@ -1,3 +1,5 @@
+from copy import deepcopy
+from functools import partial
 import json
 from unittest import TestCase
 
@@ -283,6 +285,261 @@ class TestDatasetGeo(TestCase):
             },
             {feature["id"] for feature in fc["features"]},
         )
+
+    def test_feature_to_record(self):
+        feature = {"type": "Feature", "id": "1"}
+        actual = Dataset._feature_to_record(feature, ["pk"], "geo")
+        expected = {"pk": "1"}
+        self.assertEqual(expected, actual)
+
+        feature = {
+            "type": "Feature",
+            "id": "1",
+            "geometry": {"type": "Point", "coordinates": [0, 0]},
+        }
+        actual = Dataset._feature_to_record(feature, ["pk"], "geo")
+        expected = {"pk": "1", "geo": {"point": [0, 0]}}
+        self.assertEqual(expected, actual)
+
+        feature = {
+            "type": "Feature",
+            "id": "1",
+            "geometry": {"type": "MultiPoint", "coordinates": [[0, 0], [1, 1]]},
+        }
+        actual = Dataset._feature_to_record(feature, ["pk"], "geo")
+        expected = {"pk": "1", "geo": {"multiPoint": [[0, 0], [1, 1]]}}
+        self.assertEqual(expected, actual)
+
+        feature = {
+            "type": "Feature",
+            "id": "1",
+            "geometry": {"type": "LineString", "coordinates": [[0, 0], [1, 1]]},
+        }
+        actual = Dataset._feature_to_record(feature, ["pk"], "geo")
+        expected = {"pk": "1", "geo": {"lineString": [[0, 0], [1, 1]]}}
+        self.assertEqual(expected, actual)
+
+        feature = {
+            "type": "Feature",
+            "id": "1",
+            "geometry": {
+                "type": "MultiLineString",
+                "coordinates": [[[0, 0], [1, 1], [2, 2]]],
+            },
+        }
+        actual = Dataset._feature_to_record(feature, ["pk"], "geo")
+        expected = {"pk": "1", "geo": {"multiLineString": [[[0, 0], [1, 1], [2, 2]]]}}
+        self.assertEqual(expected, actual)
+
+        feature = {
+            "type": "Feature",
+            "id": "1",
+            "geometry": {"type": "Polygon", "coordinates": [[[0, 0], [1, 1], [2, 2]]]},
+        }
+        actual = Dataset._feature_to_record(feature, ["pk"], "geo")
+        expected = {"pk": "1", "geo": {"polygon": [[[0, 0], [1, 1], [2, 2]]]}}
+        self.assertEqual(expected, actual)
+
+        feature = {
+            "type": "Feature",
+            "id": "1",
+            "geometry": {
+                "type": "MultiPolygon",
+                "coordinates": [[[[0, 0], [1, 1], [2, 2]]]],
+            },
+        }
+        actual = Dataset._feature_to_record(feature, ["pk"], "geo")
+        expected = {"pk": "1", "geo": {"multiPolygon": [[[[0, 0], [1, 1], [2, 2]]]]}}
+        self.assertEqual(expected, actual)
+
+        feature = {"type": "Feature", "id": "1", "geometry": None}
+        actual = Dataset._feature_to_record(feature, ["pk"], "geo")
+        expected = {"pk": "1"}
+        self.assertEqual(expected, actual)
+
+        feature = {
+            "type": "Feature",
+            "id": "1",
+            "bbox": [0, 0, 1, 1],
+            "geometry": {"type": "Point", "coordinates": [0, 0]},
+        }
+        actual = Dataset._feature_to_record(feature, ["pk"], "geo")
+        expected = {"pk": "1", "geo": {"point": [0, 0]}, "bbox": [0, 0, 1, 1]}
+        self.assertEqual(expected, actual)
+
+        feature = {
+            "type": "Feature",
+            "id": "1",
+            "bbox": None,
+            "geometry": {"type": "Point", "coordinates": [0, 0]},
+        }
+        actual = Dataset._feature_to_record(feature, ["pk"], "geo")
+        expected = {"pk": "1", "geo": {"point": [0, 0]}}
+        self.assertEqual(expected, actual)
+
+        feature = {
+            "type": "Feature",
+            "id": "1",
+            "bbox": [0, 0, 1, 1],
+            "geometry": {"type": "Point", "coordinates": [0, 0]},
+            "properties": {"prop1": "val1", "prop2": "val2"},
+        }
+        actual = Dataset._feature_to_record(feature, ["pk"], "geo")
+        expected = {
+            "pk": "1",
+            "geo": {"point": [0, 0]},
+            "bbox": [0, 0, 1, 1],
+            "prop1": "val1",
+            "prop2": "val2",
+        }
+        self.assertEqual(expected, actual)
+
+        feature = {
+            "type": "Feature",
+            "id": "1",
+            "bbox": [0, 0, 1, 1],
+            "geometry": {"type": "Point", "coordinates": [0, 0]},
+            "properties": None,
+        }
+        actual = Dataset._feature_to_record(feature, ["pk"], "geo")
+        expected = {"pk": "1", "geo": {"point": [0, 0]}, "bbox": [0, 0, 1, 1]}
+        self.assertEqual(expected, actual)
+
+        feature = {
+            "type": "Feature",
+            "id": "1",
+            "bbox": [0, 0, 1, 1],
+            "geometry": {"type": "Point", "coordinates": [0, 0]},
+            # Properties with names that conflict with
+            # the props in the key or geometry
+            # get ignored
+            "properties": {"pk": "val1", "geo": "val2", "bbox": "val3"},
+        }
+        actual = Dataset._feature_to_record(feature, ["pk"], "geo")
+        expected = {"pk": "1", "geo": {"point": [0, 0]}, "bbox": [0, 0, 1, 1]}
+        self.assertEqual(expected, actual)
+
+        feature = {
+            "type": "Feature",
+            "id": ["1", "2"],
+            "geometry": {"type": "Point", "coordinates": [0, 0]},
+        }
+        actual = Dataset._feature_to_record(feature, ["pk1", "pk2"], "geo")
+        expected = {"pk1": "1", "pk2": "2", "geo": {"point": [0, 0]}}
+        self.assertEqual(expected, actual)
+
+        class NotAFeature:
+            @property
+            def __geo_interface__(self):
+                return {
+                    "type": "Feature",
+                    "id": "1",
+                    "geometry": {"type": "Point", "coordinates": [0, 0]},
+                }
+
+        naf = NotAFeature()
+        actual = Dataset._feature_to_record(naf, ["pk"], "geo")
+        expected = {"pk": "1", "geo": {"point": [0, 0]}}
+        self.assertEqual(expected, actual)
+
+    @responses.activate
+    def test_from_geo_features(self):
+        def update_callback(request, snoop):
+            snoop["payload"] = request.body
+            return 200, {}, "{}"
+
+        dataset_url = f"http://localhost:9100/api/versioned/v1/datasets/1"
+        responses.add(responses.GET, dataset_url, json=self._dataset_json)
+
+        attributes_url = f"{dataset_url}/attributes"
+        responses.add(responses.GET, attributes_url, json=self._attributes_json)
+
+        records_url = f"{dataset_url}:updateRecords"
+        snoop = {}
+        responses.add_callback(
+            responses.POST, records_url, callback=partial(update_callback, snoop=snoop)
+        )
+
+        dataset = self.unify.datasets.by_resource_id("1")
+        features = [
+            {"id": "1", "geometry": {"type": "Point", "coordinates": [0, 0]}},
+            {"id": "2", "geometry": {"type": "Point", "coordinates": [1, 1]}},
+        ]
+        dataset.from_geo_features(features)
+        updates = [
+            {
+                "action": "CREATE",
+                "recordId": "1",
+                "record": {"geom": {"point": [0, 0]}, "id": "1"},
+            },
+            {
+                "action": "CREATE",
+                "recordId": "2",
+                "record": {"geom": {"point": [1, 1]}, "id": "2"},
+            },
+        ]
+        expected = "\n".join(json.dumps(update) for update in updates)
+        actual = snoop["payload"]
+        self.assertEqual(expected, actual)
+
+        class NotAFeatureCollection:
+            @property
+            def __geo_interface__(self):
+                return {"type": "FeatureCollection", "features": features}
+
+        snoop["payload"] = None
+        nafc = NotAFeatureCollection()
+        dataset.from_geo_features(nafc)
+        actual = snoop["payload"]
+        self.assertEqual(expected, actual)
+
+    @responses.activate
+    def test_from_geo_features_composite_key(self):
+        def update_callback(request, snoop):
+            snoop["payload"] = request.body
+            return 200, {}, "{}"
+
+        composite_key_dataset_json = deepcopy(self._dataset_json)
+        composite_key_dataset_json["keyAttributeNames"] = ["id1", "id2"]
+        dataset_url = f"http://localhost:9100/api/versioned/v1/datasets/1"
+        responses.add(responses.GET, dataset_url, json=composite_key_dataset_json)
+
+        composite_key_attributes_json = deepcopy(self._attributes_json)
+        composite_key_attributes_json[0]["name"] = "id1"
+        composite_key_attributes_json.insert(
+            1, deepcopy(composite_key_attributes_json[0])
+        )
+        composite_key_attributes_json[1]["name"] = "id2"
+        attributes_url = f"{dataset_url}/attributes"
+        responses.add(responses.GET, attributes_url, json=composite_key_attributes_json)
+
+        records_url = f"{dataset_url}:updateRecords"
+        snoop = {}
+        responses.add_callback(
+            responses.POST, records_url, callback=partial(update_callback, snoop=snoop)
+        )
+
+        dataset = self.unify.datasets.by_resource_id("1")
+        features = [
+            {"id": ["1", "a"], "geometry": {"type": "Point", "coordinates": [0, 0]}},
+            {"id": ["2", "b"], "geometry": {"type": "Point", "coordinates": [1, 1]}},
+        ]
+        dataset.from_geo_features(features)
+        updates = [
+            {
+                "action": "CREATE",
+                "compositeRecordId": ["1", "a"],
+                "record": {"geom": {"point": [0, 0]}, "id1": "1", "id2": "a"},
+            },
+            {
+                "action": "CREATE",
+                "compositeRecordId": ["2", "b"],
+                "record": {"geom": {"point": [1, 1]}, "id1": "2", "id2": "b"},
+            },
+        ]
+        expected = "\n".join(json.dumps(update) for update in updates)
+        actual = snoop["payload"]
+        self.assertEqual(expected, actual)
 
     _dataset_json = {
         "id": "unify://unified-data/v1/datasets/1",